// src/Models/DownloadEmandate/DownloadEmandate.ts
export interface Bankdetails {

    name: string;
    sponsorbankcode: string;

}

export interface DownloadEmandateclass {

    DebtorName: string;
    InstructedAgentMemberName : string;
    ConsumerReferenceNumber: string;
    DebtorAccountNumber: string;
    DebtorAgentID: string;
    Collection: string;
    MaximumAmount: string;
    ServiceProvider: string;
    Frequency: string;
    DebtorAccountType: string;
    CreationDateTime: string;
    length: any;

}